feat: 입력 받은 금액 동전 생성 기능 구현
