--- src/main/java/vendingmachine/domain/enums/Coin.java
@@ -1,18 +1,0 @@
-package vendingmachine.domain.enums;
-
-public enum Coin {
-    COIN_500(500),
-    COIN_100(100),
-    COIN_50(50),
-    COIN_10(10);
-
-    private final int amount;
-
-    Coin(final int amount) {
-        this.amount = amount;
-    }
-
-    public int getValue() {
-        return amount;
-    }
-}
